requirements.txt
fastapi[standard]>=0.115.12
structlog>=25.3.0
python-dotenv>=1.1.0
mcp>=1.8.0